Made reloading more efficient and fixed a bug
This commit fixes most notably the bug that if you create a project it doesnt get sorted(still doesnt) and makes it that you will still get spawned directly in the correct project when reopening the app and leaving a project alone for a while! It also makes project refresh extremely more efficient than before.