Many Improvements
This commit adds a handful of new things and a handful of fixes and improvements to libdycall for example and to the Bootstrapper aswell as some bug fixes. libdycall is rn in high development and tries to intigrate Duys LiveContainer researched techniques, give it some time and it eventually will reach its full potential later, tho cannot promise that.