fork() fix. deadlocking potential cleared up.
It doesnt matter how many times we suspend the thread and where, it would even higthen the potential a deadlock can happen, if the callers thread didnt reach suspension and the helper thread did and it continues and then resumes twice, one will KERN_SUCCESS, the other will fail, because the count is at 1. Then when the callers thread gets resumed... ur screwd because this means its incrementing the suspension count again by 1.